llm_project_env/requirements.txt
langchain
streamlit
transformers
faiss-cpu
sentence-transformers
torch
torchvision
torchaudio
requests
beautifulsoup4
